MCP-Local/requirements.txt
psycopg2
python-dotenv
fastapi
uvicorn
ollama
mcp